post authenticate delete done

+++ controllers/api/v1/posts_api.ts
@@ -13,30 +13,29 @@ export const postApiController = async (req: any, res: any) => {
     });
   return res.json(200, {
     message: "list of post",
-    post: [],
+    post: posts,
   });
 };
 
-export const destroyPost = async (req: Request, res: Response) => {
+export const destroyPost = async (req: any, res: Response) => {
   try {
-    let post: any = await Post.findById(req.params.id);
-
-    // if (post.user == req.user.id){
-    if (post) {
-      post.remove();
-
-      await Comment.deleteMany({ post: req.params.id });
-
-      return res.status(200).json({
-        message: "Post and associated comments deleted successfully!",
-      });
-      // }else{
-      //     req.flash('error', 'You cannot delete this post!');
-      //     return res.redirect('back');
-      // }
+    let post = await Post.findById(req.params.id);
+    if (post?.user?._id == req.user.id) {
+      if (post) {
+        post.deleteOne();
+        await Comment.deleteMany({ post: req.params.id });
+        return res.status(200).json({
+          message: "Post and associated comments deleted successfully!",
+        });
+      } else {
+        res.status(401).json({
+          message: "You cannot delete this post",
+        });
+      }
     }
   } catch (err) {
-    console.log("********", err);
+    console.log(err, "err");
+
     return res.status(500).json({
       message: "Internal Server Error",
     });
